requirements.txt
apache-airflow-providers-amazon==9.8.0
apache-airflow-providers-apache-spark==5.3.0
apache-airflow-providers-common-compat==1.7.0
beautifulsoup4==4.13.4
boto3==1.38.9
duckdb==1.2.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.12.0
selenium==4.32.0
webdriver-manager==4.0.2
xlrd==2.0.1